# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Random Anime Image API (pixivcat-backend)
  version: 0.0.0
  description: |
    本 OpenAPI 文档用于描述本项目新增的随机图片 API 与管理端点：
      - /random（image/json/redirect）
      - /tags（标签检索）
      - /authors（作者检索）
      - /images（列表检索）
      - /images/:id（JSON 元信息）
      - /i/:id.:ext（稳定图片 URL 的流式代理）
      - /healthz
      - /metrics
      - /admin（AdminJS + 管理 API）

    说明：
    - legacy Pixivcat 兼容路由（/:illustId.:ext 与 /:illustId-:page.:ext）在 MVP 阶段仍保留，但不纳入本 OpenAPI 的首要范围。
servers:
  - url: http://127.0.0.1:3000
...
paths:
...
  /images:
    get:
      tags: [Classification]
      summary: List images with filters and cursor pagination
      parameters:
        - $ref: '#/components/parameters/ImagesLimit'
        - $ref: '#/components/parameters/ImagesCursor'
        - $ref: '#/components/parameters/FilterR18'
        - $ref: '#/components/parameters/FilterR18Strict'
        - $ref: '#/components/parameters/FilterOrientation'
        - $ref: '#/components/parameters/FilterMinWidth'
        - $ref: '#/components/parameters/FilterMinHeight'
        - $ref: '#/components/parameters/FilterMinPixels'
        - $ref: '#/components/parameters/FilterIncludedTags'
        - $ref: '#/components/parameters/FilterExcludedTags'
        - $ref: '#/components/parameters/FilterUserId'
        - $ref: '#/components/parameters/FilterIllustId'
      responses:
        '200':
          description: Image list result
          headers:
            Cache-Control:
              schema: { type: string }
              description: Always `no-store` for this endpoint
            X-Request-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageListResponse'
        '400':
          $ref: '#/components/responses/JsonBadRequest'
        '500':
          description: Internal Server Error
          headers:
            X-Request-Id:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /images/{id}:
    get:
      tags: [Images]
      summary: Get one image record (JSON)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Image record
          headers:
            Cache-Control:
              schema: { type: string }
              description: Always `no-store` for this endpoint
            X-Request-Id:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageRecord'
        '400':
          $ref: '#/components/responses/JsonBadRequest'
        '404':
          $ref: '#/components/responses/JsonNotFound'
        '500':
          description: Internal Server Error
          headers:
            X-Request-Id:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    AdminTokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'Use: Authorization: Bearer <ADMIN_TOKEN>'
  responses:
    JsonBadRequest:
      description: Bad Request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    JsonNotFound:
      description: Not Found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  parameters:
...
    ImagesLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: /images 每页条数，默认 50，最大 200。
      examples:
        default:
          value: 50
        max:
          value: 200
...
    ImagesCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        format: int64
      description: /images 游标（上一页最后一条 image.id）。
      examples:
        next_page:
          value: 778899
    FilterR18:
      name: r18
      in: query
      required: false
      schema:
        type: integer
        enum: [0, 1, 2]
        default: 0
      description: |
        Pixiv xRestrict 过滤：
        - `0`：全年龄（默认）
        - `1`：R18
        - `2`：R18G
        仅接受 `0/1/2`；空值或其他值返回 400（code=BAD_REQUEST, message=Invalid r18.）。
      examples:
        all_ages:
          value: 0
        r18:
          value: 1
        r18g:
          value: 2
    FilterR18Strict:
      name: r18_strict
      in: query
      required: false
      schema:
        type: integer
        enum: [0, 1]
      description: |
        当 `r18=0` 时是否严格排除 `xRestrict IS NULL` 的记录。
        - `1`：严格模式（仅保留 xRestrict=0）
        - `0`：允许 xRestrict 为空的历史数据
        未传时使用服务端默认配置。
      examples:
        strict:
          value: 1
        non_strict:
          value: 0
    FilterOrientation:
      name: orientation
      in: query
      required: false
      schema:
        type: string
        enum: [portrait, landscape, square, any]
        default: any
      description: |
        图片方向筛选：
        - `any`（默认）：不筛选
        - `portrait`：竖图
        - `landscape`：横图
        - `square`：正方形
        空值或其他值返回 400（code=BAD_REQUEST, message=Invalid orientation.）。
      examples:
        any:
          value: any
        portrait:
          value: portrait
        landscape:
          value: landscape
        square:
          value: square
    FilterMinWidth:
      name: min_width
      in: query
      required: false
      schema: { type: integer, minimum: 0 }
      description: |
        最小宽度过滤（像素）。
        - 省略：不筛选
        - 仅接受非负整数；空值或非法值返回 400（code=BAD_REQUEST, message=Invalid min_width.）
      examples:
        none:
          value: 0
        hd:
          value: 1280
    FilterMinHeight:
      name: min_height
      in: query
      required: false
      schema: { type: integer, minimum: 0 }
      description: |
        最小高度过滤（像素）。
        - 省略：不筛选
        - 仅接受非负整数；空值或非法值返回 400（code=BAD_REQUEST, message=Invalid min_height.）
      examples:
        none:
          value: 0
        hd:
          value: 720
    FilterMinPixels:
      name: min_pixels
      in: query
      required: false
      schema: { type: integer, minimum: 0 }
      description: |
        最小像素数过滤：`width * height >= min_pixels`。
        - 省略：不筛选
        - 仅接受非负整数；空值或非法值返回 400（code=BAD_REQUEST, message=Invalid min_pixels.）
        - 需要图片已补全宽高元信息；否则可能被过滤掉
      examples:
        none:
          value: 0
        fullhd:
          value: 2073600
    FilterIncludedTags:
      name: included_tags
      in: query
      required: false
      schema:
        type: string
      description: |
        包含标签过滤（AND 语义）。
        - 支持多值输入（最终会合并、trim、去重）：
          - `included_tags=tag1|tag2|tag3`（`|` 分隔）
          - `included_tags=tag1,tag2,tag3`（`,` 分隔）
          - `included_tags=tag1&included_tags=tag2`（重复 query param）
        - 语义：必须同时包含所有给定标签（AND）
        空值或解析后为空将返回 400（code=BAD_REQUEST, message=Invalid included_tags.）。
      examples:
        single:
          value: cat
        and_semantics:
          value: cat|dog
    FilterExcludedTags:
      name: excluded_tags
      in: query
      required: false
      schema:
        type: string
      description: |
        排除标签过滤（NOT 语义）。
        - 支持多值输入（最终会合并、trim、去重）：
          - `excluded_tags=tag1|tag2|tag3`（`|` 分隔）
          - `excluded_tags=tag1,tag2,tag3`（`,` 分隔）
          - `excluded_tags=tag1&excluded_tags=tag2`（重复 query param）
        - 语义：只要命中任意一个排除标签就会被过滤掉（NOT / OR）
        空值或解析后为空将返回 400（code=BAD_REQUEST, message=Invalid excluded_tags.）。
      examples:
        single:
          value: r18
        not_semantics:
          value: r18|r18g
    FilterUserId:
      name: user_id
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
      description: |
        Pixiv 用户 ID 过滤（精确匹配）。
        - 省略：不筛选
        - 仅接受正整数；空值或非法值返回 400（code=BAD_REQUEST, message=Invalid user_id.）
      examples:
        example:
          value: 12345678
    FilterIllustId:
      name: illust_id
      in: query
      required: false
      schema:
        type: integer
        format: int64
        minimum: 1
      description: |
        Pixiv 作品 ID 过滤（精确匹配）。
        - 省略：不筛选
        - 仅接受正整数；空值或非法值返回 400（code=BAD_REQUEST, message=Invalid illust_id.）
      examples:
        example:
          value: 123456789
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          examples: [NO_MATCH, UNSUPPORTED_URL, UPSTREAM_RATE_LIMIT]
        message:
          type: string
        request_id:
          type: string
...
    ImageListResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, illust_id, page_index, ext, width, height, x_restrict, user_id, user_name, status, tags]
            properties:
              id: { type: integer }
              illust_id: { type: integer }
              page_index: { type: integer, minimum: 0 }
              ext:
                type: string
                enum: [jpg, jpeg, png, gif]
              width: { type: integer, nullable: true }
              height: { type: integer, nullable: true }
              x_restrict:
                type: integer
                nullable: true
                description: 'Pixiv restriction: 0/1/2'
              user_id: { type: integer, nullable: true }
              user_name: { type: string, nullable: true }
              status:
                type: string
                enum: [active, disabled, broken]
              tags:
                type: array
                items: { type: string }
        next_cursor:
          type: string
          nullable: true
...
    ImageRecord:
      type: object
      required: [id, illust_id, page_index]
      properties:
        id: { type: integer }
        illust_id: { type: integer }
        page_index: { type: integer, minimum: 0 }
        ext:
          type: string
          enum: [jpg, jpeg, png, gif]
        width: { type: integer, nullable: true }
        height: { type: integer, nullable: true }
        x_restrict:
          type: integer
          nullable: true
          description: 'Pixiv restriction: 0/1/2'
        user_id: { type: integer, nullable: true }
        user_name: { type: string, nullable: true }
        status:
          type: string
          enum: [active, disabled, broken]
        tags:
          type: array
          items: { type: string }
        fail_count: { type: integer, nullable: true }
        last_fail_at:
          type: string
          format: date-time
          nullable: true
...